fix: chunked loading
