+++ package.json
@@ -74,8 +74,7 @@
     "slate": "^0.73.1",
     "slate-history": "^0.66.0",
     "slate-react": "^0.74.2",
-    "source-map-explorer": "^2.5.3",
-    "tinycolor2": "^1.4.2"
+    "source-map-explorer": "^2.5.3"
   },
   "scripts": {
     "webpack": "webpack",
