+++ src/App.tsx
@@ -1,13 +1,5 @@
-
 import Main from "./views/Main";
 
 export default function App() {
-//    const url = useSelector((store:storeProps) => store.apiUrl)
-//     useEffect(()=>{
-//         console.log('making apiUrl request to',url)
-//         boscoRequest('https://cdn.jsdelivr.net/gh/metrico/boscaiolog@main/boscaiolog.js',{url})
-//     },[])
-
-
     return <Main />;
 }
