+++ src/components/DataViews/components/QueryBuilder/LabelValueForm.tsx
@@ -43,7 +43,7 @@ export const LabelValueForm = (props: any) => {
     const onLabelChange = (e: any) => {
         const { value, id } = e;
 
-        setLabelValue((_: any) => ({
+        setLabelValue(() => ({
             value: value?.value,
             label: value?.value,
         }));
@@ -81,7 +81,7 @@ export const LabelValueForm = (props: any) => {
         const prevKeyVal = JSON.parse(JSON.stringify(localKeyVal));
         const newKeyVal = { ...prevKeyVal, operator: value?.value, id };
         onChange(newKeyVal);
-        setOperatorValue((_: any) => {
+        setOperatorValue(() => {
             return { ...value };
         });
         setLocalKeyVal((prev: any) => ({ ...prev, operator: value?.value }));
