+++ src/components/DataViews/views/LogsView.tsx
@@ -57,7 +57,7 @@ export default function LogsView(props: any) {
         };
         localStorage.setItem("tabsState", JSON.stringify(newState));
         if (typeof value === "number") {
-            setTabsState((_: any) => value);
+            setTabsState(() => value);
         }
     };
     return (
