+++ src/components/QueryBuilder/Operations/consts.tsx
@@ -1,0 +1,32 @@
+export const ranges = [
+    "rate",
+    "rate_counter",
+    "count_over_time",
+    "sum_over_time",
+    "bytes_rate",
+    "bytes_over_time",
+    "absent_over_time",
+];
+
+export const label_ranges = [
+    "avg_over_time",
+    "max_over_time",
+    "min_over_time",
+    "first_over_time",
+    "last_over_time",
+    "stdvar_over_time",
+    "stddev_over_time",
+    "quantile_over_time",
+    "",
+];
+
+export const aggregations = [
+    "sum",
+    "min",
+    "max",
+    "avg",
+    "stddev",
+    "stdvar",
+    "count",
+];
+export const aggregations_k = ["topk", "bottomk"];
