+++ src/components/QueryTypeBar/index.tsx
@@ -14,11 +14,11 @@ import { useTheme } from "../DataViews/components/QueryBuilder/hooks";
 const QueryTypeCont = styled.div`
     display: flex;
     padding: 4px;
-    background: ${(props:any) => props.theme.widgetContainer};
-    color: ${(props:any) => props.color};
+    background: ${(props: any) => props.theme.widgetContainer};
+    color: ${(props: any) => props.color};
     height: 26px;
 `;
-export function panelAction(name:any, value:any) {
+export function panelAction(name: any, value: any) {
     if (name === "left") {
         return setLeftPanel(value);
     }
@@ -35,14 +35,14 @@ export const DIRECTION_SWITCH_OPTIONS = [
     { value: "backwards", label: "Backwards" },
 ];
 
-export default function QueryTypeBar(props:any) {
+export default function QueryTypeBar(props: any) {
     const dispatch = useDispatch();
     const { name, data } = props;
 
-  const theme = useTheme()
-    const panelQuery = useSelector((store:any) => store[name]);
+    const theme = useTheme();
+    const panelQuery = useSelector((store: any) => store[name]);
 
-    const responseType = useSelector((store:any) => store.responseType);
+    const responseType = useSelector((store: any) => store.responseType);
 
     const { hash } = useLocation();
     const {
@@ -62,7 +62,9 @@ export default function QueryTypeBar(props:any) {
     const [isTableViewSet, setIsTableViewSet] = useState(tableView);
     const [isShowTsSet, setIsShowTsSet] = useState(isShowTs || false);
     const [isBuilderSet, setIsBuilderSet] = useState(isBuilder || false);
-    const [isLogsVolumeSet, setIsLogsVolumeSet] = useState( isLogsVolume || false);
+    const [isLogsVolumeSet, setIsLogsVolumeSet] = useState(
+        isLogsVolume || false
+    );
     const [isShowStatsSet, setIsShowStatsSet] = useState(isShowStats || false);
     const [queryTypeSwitch, setQueryTypeSwitch] = useState(queryType);
     const [directionSwitch, setDirectionSwitch] = useState(
@@ -71,12 +73,12 @@ export default function QueryTypeBar(props:any) {
 
     useEffect(() => {
         const urlParams = new URLSearchParams(hash.replace(/#/, ""));
-        const urlPanel:any = urlParams.get(name);
+        const urlPanel: any = urlParams.get(name);
 
-        const parsedPanel = JSON.parse(decodeURIComponent(urlPanel)||'[]');
+        const parsedPanel = JSON.parse(decodeURIComponent(urlPanel) || "[]");
 
         if (parsedPanel?.length > 0) {
-            const queryMD = parsedPanel.find((f:any) => f.idRef === idRef);
+            const queryMD = parsedPanel.find((f: any) => f.idRef === idRef);
 
             if (queryMD) {
                 const panel = [...panelQuery];
@@ -89,7 +91,7 @@ export default function QueryTypeBar(props:any) {
                 dispatch(panelAction(name, panel));
             }
         }
-          // eslint-disable-next-line react-hooks/exhaustive-deps
+        // eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
 
     useEffect(() => {
@@ -103,11 +105,11 @@ export default function QueryTypeBar(props:any) {
         setIsBuilderSet(props.data.isBuilder);
     }, [setIsBuilderSet, props.data.isBuilder]);
 
-    useEffect(()=>{
+    useEffect(() => {
         setIsLogsVolumeSet(props.data.isLogsVolume);
-    },[setIsLogsVolumeSet, props.data.isLogsVolume])
+    }, [setIsLogsVolumeSet, props.data.isLogsVolume]);
 
-    function onSwitchChange(e:any) {
+    function onSwitchChange(e: any) {
         // modify query type switch value
         const panel = [...panelQuery];
         panel.forEach((query) => {
@@ -119,7 +121,7 @@ export default function QueryTypeBar(props:any) {
         setQueryTypeSwitch(e);
     }
 
-    function onDirectionSwitchChange(e:any) {
+    function onDirectionSwitchChange(e: any) {
         // modify query type switch value
         const panel = [...panelQuery];
         panel.forEach((query) => {
@@ -168,18 +170,18 @@ export default function QueryTypeBar(props:any) {
 
     function handleLogsVolumeSwitch() {
         const panel = [...panelQuery];
-        panel.forEach((query)=> {
-            if(query.id === id) {
-                query.isLogsVolume = isLogsVolumeSet ? false : true
+        panel.forEach((query) => {
+            if (query.id === id) {
+                query.isLogsVolume = isLogsVolumeSet ? false : true;
             }
-        })
+        });
 
         dispatch(panelAction(name, panel));
     }
 
-    function handleStatsInfoSwitch(e:any) {
+    function handleStatsInfoSwitch(e: any) {
         const value = e.target.checked;
-        setIsShowStatsSet((_:any) => value);
+        setIsShowStatsSet(() => value);
         const panel = [...panelQuery];
         panel.forEach((query) => {
             if (query.id === id) {
