+++ src/views/DataSources/components/DataSourceHeaders.tsx
@@ -7,19 +7,19 @@ import { InputCol, InputGroup } from "../styles";
 import { Field } from "../ui";
 import { SectionHeader } from "./SectionHeader";
 
-export const DataSourceHeaders = (props:any) => {
+export const DataSourceHeaders = (props: any) => {
     const dispatch = useDispatch();
 
-    const dataSources = useSelector((store:any) => store.dataSources);
-    const [editing, setEditing] = useState(false)
+    const dataSources = useSelector((store: any) => store.dataSources);
+    const [editing, setEditing] = useState(false);
     const { headers, id } = props;
-    const onChange = (e:any, headerId:any, name:any) => {
-        setEditing((_)=> true)
+    const onChange = (e: any, headerId: any, name: any) => {
+        setEditing(() => true);
         const value = e.target.value; // identify value changed
-        const prevDs = dataSources.find((f:any) => f.id === id);
+        const prevDs = dataSources.find((f: any) => f.id === id);
         const prevHeaders = prevDs["headers"] || [];
 
-        const newHeaders = prevHeaders.map((m:any) => {
+        const newHeaders = prevHeaders.map((m: any) => {
             if (m.id === headerId) {
                 m[name] = value;
                 return m;
@@ -27,7 +27,7 @@ export const DataSourceHeaders = (props:any) => {
             return m;
         });
 
-        const newDataSources = dataSources.map((ds:any) => {
+        const newDataSources = dataSources.map((ds: any) => {
             if (ds.id === id) {
                 ds.headers = [...newHeaders];
             }
@@ -37,13 +37,13 @@ export const DataSourceHeaders = (props:any) => {
         localStorage.setItem("dataSources", JSON.stringify(newDataSources));
         dispatch(setDataSources(newDataSources));
         setTimeout(() => {
-            setEditing((_) => false);
+            setEditing(() => false);
         }, 800);
     };
 
-    const onAdd = (e:any) => {
+    const onAdd = (e: any) => {
         e.preventDefault();
-        const prevDataSource = dataSources?.find((f:any) => f.id === id);
+        const prevDataSource = dataSources?.find((f: any) => f.id === id);
         const prevHeaders = prevDataSource["headers"];
         const dsClone = JSON.parse(JSON.stringify([...dataSources]));
 
@@ -59,7 +59,7 @@ export const DataSourceHeaders = (props:any) => {
                 headers: [...prevDataSource.headers, newHeaders],
             };
 
-            const newDataSources = dsClone?.map((ds:any) => {
+            const newDataSources = dsClone?.map((ds: any) => {
                 if (ds.id === id) {
                     return { ...newDataSource };
                 }
@@ -71,12 +71,12 @@ export const DataSourceHeaders = (props:any) => {
         }
     };
 
-    const onRemove = (_:any, headerId:any) => {
-        const prevDataSource = dataSources?.find((f:any) => f.id === id);
+    const onRemove = (_: any, headerId: any) => {
+        const prevDataSource = dataSources?.find((f: any) => f.id === id);
         const prevHeaders = [...headers];
         const headerRm = prevHeaders.filter((f) => f.id !== headerId);
         const newDataSource = { ...prevDataSource, headers: [...headerRm] };
-        const newDataSources = dataSources?.map((ds:any) => {
+        const newDataSources = dataSources?.map((ds: any) => {
             if (ds.id === id) {
                 return { ...newDataSource };
             }
@@ -85,7 +85,6 @@ export const DataSourceHeaders = (props:any) => {
 
         localStorage.setItem("dataSources", JSON.stringify(newDataSources));
         dispatch(setDataSources(newDataSources));
-
     };
 
     return (
@@ -100,26 +99,29 @@ export const DataSourceHeaders = (props:any) => {
                         onClickAdd={onAdd}
                     />
 
-                    {headers?.map((val:any, key:any) => (
+                    {headers?.map((val: any, key: any) => (
                         <InputCol key={key}>
                             <InputGroup>
                                 <Field
                                     label={"header"}
                                     value={val.header}
-                                    onChange={(e:any) =>
+                                    onChange={(e: any) =>
                                         onChange(e, val.id, "header")
                                     }
                                 />
                                 <Field
                                     label={"value"}
                                     value={val.value}
-                                    onChange={(e:any) =>
+                                    onChange={(e: any) =>
                                         onChange(e, val.id, "value")
                                     }
                                 />
                                 <DeleteOutlineOutlinedIcon
                                     onClick={(e) => onRemove(e, val.id)}
-                                    style={{ cursor: "pointer" , marginLeft:'10px'}}
+                                    style={{
+                                        cursor: "pointer",
+                                        marginLeft: "10px",
+                                    }}
                                     fontSize={"small"}
                                 />
                             </InputGroup>
