+++ src/views/DataSources/components/DataSourcesFiller.tsx
@@ -6,7 +6,7 @@ import setDataSources from "../store/setDataSources";
 import { Button, Field } from "../ui";
 import { themes } from "../../../components/DataViews/components/Traces/Jaeger-ui/src/theme/themes";
 
-const InlineFlex = (theme:any) => css`
+const InlineFlex = (theme: any) => css`
     display: flex;
     flex-direction: column;
     flex: 1;
@@ -64,29 +64,30 @@ export const DataSourcesFiller = (props: any) => {
     const dispatch = useDispatch();
     const submitMessage = "Save";
 
-    const storeTheme = useSelector((store:{theme:'light'|'dark'})=> store.theme)
+    const storeTheme = useSelector(
+        (store: { theme: "light" | "dark" }) => store.theme
+    );
 
-    const theme = useMemo (()=>{
-        return themes[storeTheme]
-
-    },[storeTheme])
+    const theme = useMemo(() => {
+        return themes[storeTheme];
+    }, [storeTheme]);
 
     const urlChange = (e: any) => {
-        setUrl((_) => e.target.value);
+        setUrl(() => e.target.value);
     };
     const userChange = (e: any) => {
-        setUser((_) => e.target.value);
+        setUser(() => e.target.value);
     };
     const passwordChange = (e: any) => {
-        setPassword((_) => e.target.value);
+        setPassword(() => e.target.value);
     };
 
     const onSwitchChange = (e: any) => {
-        setOneForAll((_) => e.target.checked);
+        setOneForAll(() => e.target.checked);
     };
 
     const onBasicAuthChange = (e: any) => {
-        setBasicAuth((_) => e.target.checked);
+        setBasicAuth(() => e.target.checked);
     };
 
     const onUseForAll = (e: any) => {
